clean up a few unused files
