+++ src/calicost/utils_hmrf.py
@@ -408,61 +408,3 @@ def load_hmrf_given_iteration(filename, r):
         res["barcodes"] = allres["barcodes"]
     return res
 
-
-def get_LoH_for_phylogeny(df_seglevel_cnv, min_num_bins=3):
-    """
-    Treating LoH as irreversible point mutations, output a clone-by-mutation matrix for phylogeny reconstruction.
-    Mutation states: 0 for no LoH, 1 for lossing A allele, 2 for lossing B allele.
-
-    Attributes
-    ----------
-    df_seglevel_cnv : pd.DataFrame, (n_obs, 3+2*n_clones)
-        Dataframe from cnv_*seglevel.tsv output.
-
-    Returns
-    ----------
-    df_loh : pd.DataFrame, (n_clones, n_segments)
-    """
-    def get_shared_intervals(acn_profile):
-        '''
-        Takes in allele-specific copy numbers, output a segmentation of genome such that all clones are in the same CN state within each segment.
-
-        anc_profile : array, (n_obs, 2*n_clones)
-            Allele-specific integer copy numbers for each genomic bin (obs) across all clones.
-        '''
-        intervals = []
-        seg_acn = []
-        s = 0
-        while s < acn_profile.shape[0]:
-            t = np.where( ~np.all(acn_profile[s:,] == acn_profile[s,:], axis=1) )[0]
-            if len(t) == 0:
-                intervals.append( (s, acn_profile.shape[0])  )
-                seg_acn.append( acn_profile[s,:] )
-                s = acn_profile.shape[0]
-            else:
-                t = t[0]
-                intervals.append( (s,s+t) )
-                seg_acn.append( acn_profile[s,:] )
-                s = s+t
-        return intervals, seg_acn
-    
-    clone_ids = [x.split(" ")[0] for x in df_seglevel_cnv.columns[ np.arange(3, df_seglevel_cnv.shape[1], 2) ] ]
-    
-    acn_profile = df_seglevel_cnv.iloc[:,3:].values
-    intervals, seg_acn = get_shared_intervals(acn_profile)
-    df_loh = []
-    for i, acn in enumerate(seg_acn):
-        if np.all(acn != 0):
-            continue
-        if intervals[i][1] - intervals[i][0] < min_num_bins:
-            continue
-        idx_zero = np.where(acn == 0)[0]
-        idx_clones = (idx_zero / 2).astype(int)
-        is_A = (idx_zero % 2 == 0)
-        # vector of mutation states
-        mut = np.zeros( int(len(acn) / 2), dtype=int )
-        mut[idx_clones] = np.where(is_A, 1, 2)
-        df_loh.append( pd.DataFrame(mut.reshape(1, -1), index=[f"bin_{intervals[i][0]}_{intervals[i][1]}"], columns=clone_ids) )
-
-    df_loh = pd.concat(df_loh).T
-    return df_loh
